+++ packages/nesd/test/ui/settings/settings_screen_test.dart
@@ -1,6 +1,7 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:nesd/ui/common/settings_tile.dart';
 import 'package:nesd/ui/settings/audio/low_pass_filter_switch.dart';
+import 'package:nesd/ui/settings/audio/swap_duty_cycles_switch.dart';
 import 'package:nesd/ui/settings/debug/debug_overlay_switch.dart';
 import 'package:nesd/ui/settings/debug/debug_settings.dart';
 
@@ -50,6 +51,11 @@ void main() {
     await r.expectSwitch(
       find.byType(LowPassFilterSwitch),
       getValue: () => r.settings.lowPassFilter,
+    );
+
+    await r.expectSwitch(
+      find.byType(SwapDutyCyclesSwitch),
+      getValue: () => r.settings.swapDutyCycles,
     );
   });